Visualizer seems to be working

+++ tools/build_report.py
@@ -1,7 +1,7 @@
 from typing import *
 
 class BuildReport:
-    def __init__(self, dataset: str, format: str, time: int,
+    def __init__(self, dataset: str, format: str, time: float,
                  created_files: List[str]):
         self.dataset = dataset
         self.format = format
